Refactor vote modal
