+++ pages/api/comments/executive/add/[address].ts
@@ -6,6 +6,7 @@ import { connectToDatabase } from 'lib/api/utils';
 import withApiHandler from 'lib/api/withApiHandler';
 import { config } from 'lib/config';
 import { DEFAULT_NETWORK, SupportedNetworks } from 'lib/constants';
+import { ExecutiveComment, ExecutiveCommentsRequestBody } from 'modules/comments/types/executiveComment';
 
 export default withApiHandler(
   async (req: NextApiRequest, res: NextApiResponse) => {
@@ -16,7 +17,14 @@ export default withApiHandler(
 
     invariant(network && network.length > 0, 'Network not supported');
 
-    const { voterAddress, comment, commentSig, txHash, voterWeight } = JSON.parse(req.body);
+    const {
+      voterAddress,
+      comment,
+      signedMessage,
+      txHash,
+      voterWeight,
+      delegateAddress
+    }: ExecutiveCommentsRequestBody = JSON.parse(req.body);
 
     const provider = ethers.getDefaultProvider(SupportedNetworks.MAINNET, {
       infura: config.INFURA_KEY,
@@ -32,7 +40,7 @@ export default withApiHandler(
 
     // verify signature
     invariant(
-      ethers.utils.verifyMessage(comment, commentSig) === ethers.utils.getAddress(voterAddress),
+      ethers.utils.verifyMessage(comment, signedMessage) === ethers.utils.getAddress(voterAddress),
       'invalid message signature'
     );
 
@@ -42,14 +50,18 @@ export default withApiHandler(
     invariant(await client.isConnected(), 'Mongo client failed to connect');
 
     const collection = db.collection('executiveComments');
-    await collection.insertOne({
+
+    const newComment: ExecutiveComment = {
       spellAddress,
       voterAddress,
+      delegateAddress,
       comment,
       voterWeight,
       date: new Date(),
       network
-    });
+    };
+
+    await collection.insertOne(newComment);
 
     res.status(200).json({ success: 'Added Successfully' });
   },
